- Initial implementtion of ASP-style rule constraints. Cyclic rules are handled correctly. Lots of room for optimization. Graph-based rule constraint API not implemented yet. - Added better debug visualization for VarID/Literal and rule atoms/bodies. - Added KnightTour solver test, specifically to test cyclic rule handling. - Added some more basic rule tests.

+++ vertexy/src/public/ds/FastLookupSet.h
@@ -188,7 +188,7 @@ protected:
 		m_inSet[ival] = m_curStamp;
 	}
 
-	vector<uint8_t> m_inSet;
+	vector<uint32_t> m_inSet;
 	vector<T> m_elements;
 	uint32_t m_curStamp = 1;
 	mutable ToIndexType m_indexer;
